requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
langchain==0.0.137
openai==0.28